Moved MockEclairClient to testkit
